% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMayoMap.R
\name{cumvar}
\alias{cumvar}
\title{Calculate the cumulative variance of a vector}
\usage{
cumvar(x)
}
\arguments{
\item{x}{a vector}
}
\value{
{v}{a vector of cumulative variance}
}
\description{
Calculate the cumulative variance of a vector
}
