% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMayoMap.R
\docType{data}
\name{demodata}
\alias{demodata}
\title{An illustrative user-supplied marker pool
This dataset serves as a demonstration and includes a collection of user-specified marker pools.}
\format{
A data frame with 3 columns:
\describe{
  \item{gene}{Names of marker genes}
  \item{celltype}{Names of cell types}
  \item{value}{A column populated with the value of 1, signifying the presence of the corresponding marker gene for the respective cell type}
}
}
\usage{
data(demodata)
}
\description{
An illustrative user-supplied marker pool
This dataset serves as a demonstration and includes a collection of user-specified marker pools.
}
\examples{
# If 'demodata' is your marker pool, use the code below to modify the marker
# pool into the database format required by scMayoMap:
library(scMayoMap)
data(demodata)
db <- tidyr::spread(demodata, key = c('celltype'), value = 'value')
db[is.na(db)] <- 0
obj <- scMayoMap(data = data, database = db)
}
\keyword{datasets}
