% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMayoMap.R
\name{scMayoMap.plot}
\alias{scMayoMap.plot}
\title{plot the annotation result of scMayoMap}
\usage{
scMayoMap.plot(scMayoMap.object, directory = NULL, width = 8, height = 6)
}
\arguments{
\item{scMayoMap.object}{return from function \code{scMayoMap}.}

\item{directory}{character. the directory to save the figures. Default is the current working directory.}

\item{width}{the width of the graphics region in inches. See R function \code{pdf}.}

\item{height}{the height of the graphics region in inches. See R function \code{pdf}.}
}
\value{
\item{annotation.plot}{A ggplot object}
}
\description{
plot the annotation result of scMayoMap
}
