% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMayoMap.R
\name{scMayoMap}
\alias{scMayoMap}
\title{scMayoMap}
\usage{
scMayoMap(
  data,
  database = NULL,
  padj.cutoff = 0.05,
  pct.cutoff = 0.25,
  tissue = NULL
)
}
\arguments{
\item{data}{data frame. Rows: gene names; columns must include "avg_log2FC", "pct.1","pct.2", "gene".}

\item{database}{data frame. Already cleaned database saved in the package.}

\item{padj.cutoff}{number between 0-1. Used to filter the genes by FDR adjusted pvalue.}

\item{pct.cutoff}{number between 0-1. Genes with pct.1 greater than this number will be remained for later analysis.}

\item{tissue}{tissue to be specified. If NULL, tissue will not be specified. User can choose one of:
"adipose tissue"; "bladder"; "blood"; "bone"; "bone marrow";
"brain";"breast"; "embryo"; "eye"; "gastrointestinal tract";
"heart";"kidney";"liver";"lung";"mammary gland";"muscle";
"other";"ovary";"pancreas";"placenta";"prostate";
"skin";"spleen";"stomach";"testis";"thymus";"tooth";"uterus".}
}
\value{
A list with the elements
\item{res}{A data.frame listed cluster("cluster") with the top N potential celltypes("celltype") with estimated scores "ES.raw", standard errors "ES.se", normalized estimated scores "ES.norm", and markers matched for predicted cell type.}
\item{markers}{A data.frame of cell types scores of each cluster. Row: cluster, Column: annotated cell types.}
\item{tissue}{tissue specified in the analysis.}
\item{annotation.norm}{A data.frame of cell types scores of each cluster. Row: cluster, Column: annotated cell types.}
\item{annotation.mean}{A data.frame of cell types mean of each cluster. Row: cluster, Column: annotated cell types.}
}
\description{
scMayoMap
}
