% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R, R/scMayoMap.R
\docType{data}
\name{scMayoMapDatabase}
\alias{scMayoMapDatabase}
\title{database}
\format{
A \code{data.frame} with columns and rows:
\describe{
\item{tissue}{column of tissue names}
\item{gene}{column of gene names}
}

An object of class \code{data.frame} with 21658 rows and 606 columns.
}
\usage{
scMayoMapDatabase

data(scMayoMapDatabase)
}
\description{
A \code{data.frame} lists all marker genes for each cell type in a variety of tissues
}
\examples{
data(scMayoMapDatabase)
}
\keyword{datasets}
