% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\docType{data}
\name{scMayoMapDatabase}
\alias{scMayoMapDatabase}
\title{database}
\format{
A \code{data.frame} with columns and rows:
\describe{
\item{tissue}{column of tissue names}
\item{gene}{column of gene names}
}
}
\usage{
scMayoMapDatabase
}
\description{
A \code{data.frame} lists all marker genes for each cell type in a variety of tissues
}
\keyword{datasets}
